:construction: Initial implementation of conda-lock GitHub Action
Trying out a composite GitHub Action first, instead of docker or javascript one. Just two inputs for now, the file path to the environment specification file, and the platform to generate the lockfiles for. Installing conda-lock via pip from setup-python, running the conda-lock command, inspect the output and try to commit files (if allowed).