:sparkles: Initial implementation of conda-lock GitHub Action (#1)
* :construction: Initial implementation of conda-lock GitHub Action

Trying out a composite GitHub Action first, instead of docker or javascript one. Just two inputs for now, the file path to the environment specification file, and the platform to generate the lockfiles for. Installing conda-lock via pip from setup-python, running the conda-lock command, inspect the output and try to commit files (if allowed).

* :seedling: Example GitHub Actions workflow and environment.yml file

Minimal environment.yml with numpy only, and a GitHub Action workflow to test if things work.

* :bug: Set shell for commit condalock files step

To fix `Required property is missing: shell`.

* :alembic: Try git checkout with repository and ref

Explicitly checkout the current repository and branch. Also disable running this test-lock.yml workflow on pushes to main branch.

* :alembic: Try checkout github.head_ref

The workflow isn't triggered from repository_dispatch, so probably just need to grab the current branch the Pull Request is in.

* :recycle: Split git commit --all to git add -all && git commit

Newly generated conda-lock.yml files are not added with `git commit --all`, so need to split the command into `git add all && git commit`, see https://stackoverflow.com/questions/2419249/how-can-i-stage-and-commit-all-files-including-newly-added-files-using-a-singl/2419270#2419270

* :passport_control: Give workflow write permission

Needed for the default token to do `git push`.

* [condalock-command] autogenerated conda-lock files

---------

Co-authored-by: actions-bot <58130806+actions-bot@users.noreply.github.com>